+++ packages/views/auth/index.ts
@@ -1,1 +1,2 @@
 export { LoginPage, validateCliCallback } from "./login-page";
+export { useLogout } from "./use-logout";
